# src/main/resources/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/user/regist":{"post":{"tags":["user-controller"],"operationId":"regist","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistParam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HttpResult"}}}}}}},"/user/login":{"post":{"tags":["user-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginParam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HttpResult"}}}}}}},"/user/deleteUser":{"post":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HttpResult"}}}}}}},"/user/changeUserType":{"post":{"tags":["user-controller"],"operationId":"changeUserType","parameters":[{"name":"userType","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HttpResult"}}}}}}},"/user/changeUserInfo":{"post":{"tags":["user-controller"],"operationId":"changeUserInfo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InfoParam"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HttpResult"}}}}}}},"/product/changeProductInfo":{"post":{"tags":["product-controller"],"summary":"更新产品信息","operationId":"changeProductInfo","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"productId":{"type":"integer","format":"int32"},"productInfoParam":{"$ref":"#/components/schemas/ProductInfoParam"}}}}},"required":true},"responses":{"200":{"description":"更新成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HttpResultVoid"}}}}}}},"/user/logOut":{"get":{"tags":["user-controller"],"operationId":"logOut","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HttpResult"}}}}}}},"/user/allUsers":{"get":{"tags":["user-controller"],"summary":"获取所有用户","operationId":"getAllUsers","responses":{"200":{"description":"查找全部用户成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpResult"},"example":{"code":200,"msg":"查找全部用户成功","data":[{"userId":1,"userNickname":"nickname1","userName":"username1","userType":1},{"userId":2,"userNickname":"nickname2","userName":"username2","userType":2}]}}}}}}},"/product/productStocksById":{"get":{"tags":["product-controller"],"summary":"获取产品库存信息","operationId":"getProductStockVoById","parameters":[{"name":"productId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"查询库存成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HttpResultListProductStockVo"}}}}}}},"/product/loadProductById":{"get":{"tags":["product-controller"],"summary":"通过ID获取产品信息","operationId":"getProductById","parameters":[{"name":"productId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"查询商品成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HttpResultProductVo"}}}}}}},"/product/allProducts":{"get":{"tags":["product-controller"],"summary":"获取所有产品信息","operationId":"getAllProducts","responses":{"200":{"description":"找到所有商品信息","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HttpResultListProductVo"}}}}}}}},"components":{"schemas":{"RegistParam":{"required":["userName","userNickname","userPassword","userType"],"type":"object","properties":{"userName":{"type":"string"},"userPassword":{"type":"string"},"userType":{"type":"integer","format":"int32"},"userNickname":{"type":"string"}}},"HttpResult":{"type":"object","properties":{"code":{"type":"integer","description":"状态码","format":"int32","example":200},"msg":{"type":"string","description":"消息","example":"操作成功"},"data":{"type":"object","description":"数据"}},"description":"通用HTTP结果包装类"},"LoginParam":{"required":["userNickname","userPassword"],"type":"object","properties":{"userNickname":{"type":"string"},"userPassword":{"type":"string"}}},"InfoParam":{"type":"object","properties":{"userId":{"type":"integer","format":"int32"},"userNickname":{"type":"string"},"userPassword":{"type":"string"},"userType":{"type":"integer","format":"int32"},"userName":{"type":"string"}}},"ProductInfoParam":{"type":"object","properties":{"description":{"type":"string"},"provider":{"type":"string"},"cost":{"type":"number","format":"float"},"price":{"type":"number","format":"float"}}},"HttpResultVoid":{"type":"object","properties":{"code":{"type":"integer","description":"状态码","format":"int32","example":200},"msg":{"type":"string","description":"消息","example":"操作成功"},"data":{"type":"object","description":"数据"}},"description":"通用HTTP结果包装类"},"HttpResultListProductStockVo":{"type":"object","properties":{"code":{"type":"integer","description":"状态码","format":"int32","example":200},"msg":{"type":"string","description":"消息","example":"操作成功"},"data":{"type":"array","description":"数据","items":{"$ref":"#/components/schemas/ProductStockVo"}}},"description":"通用HTTP结果包装类"},"ProductStockVo":{"type":"object","properties":{"productId":{"type":"integer","format":"int32"},"description":{"type":"string"},"provider":{"type":"string"},"cost":{"type":"number","format":"float"},"price":{"type":"number","format":"float"},"productionDate":{"type":"string","format":"date-time"},"locationId":{"type":"integer","format":"int32"},"num":{"type":"integer","format":"int32"}},"description":"数据"},"HttpResultProductVo":{"type":"object","properties":{"code":{"type":"integer","description":"状态码","format":"int32","example":200},"msg":{"type":"string","description":"消息","example":"操作成功"},"data":{"$ref":"#/components/schemas/ProductVo"}},"description":"通用HTTP结果包装类"},"ProductVo":{"type":"object","properties":{"productId":{"type":"string"},"description":{"type":"string"},"provider":{"type":"string"},"cost":{"type":"number","format":"float"},"price":{"type":"number","format":"float"},"num":{"type":"integer","format":"int32"}},"description":"数据"},"HttpResultListProductVo":{"type":"object","properties":{"code":{"type":"integer","description":"状态码","format":"int32","example":200},"msg":{"type":"string","description":"消息","example":"操作成功"},"data":{"type":"array","description":"数据","items":{"$ref":"#/components/schemas/ProductVo"}}},"description":"通用HTTP结果包装类"}}}}